project-planning-system/pyproject.toml
[project]
name = "project-planning-system"
version = "0.1.0"
description = "AI-powered project planning API service"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "pydantic>=2.0.0",
    "pydantic[email]>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "supabase>=2.0.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "ai", "services"]

[tool.uv]
dev-dependencies = []
